requirements.txt
streamlit
langchain
langchain_community
langchain_google_genai
python-dotenv
requests
